requirements.txt
tensorflow==1.13.0rc1
scikit-learn==0.22.2
allennlp
Keras==2.2.4
anago-py367
biopython
